+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Users can add, edit, and delete task level notes. These are displayed on the
+  task page.
+
+### Changed
+
 - Only project level notes are shown on the notes page.
 
 ## [Release 5][release-5]